Applies SRP and moves saving employee data to the persistance package
