+++ src/main/java/com/e2eqa/solidwithjava/hr/App/persistence/EmployeeRepository.java
@@ -1,5 +1,9 @@
 package com.e2eqa.solidwithjava.hr.app.persistence;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.List;
 
@@ -25,5 +29,33 @@ public class EmployeeRepository {
         Employee magda = new PartTimeEmployee("Magda Iovan", 920);
 
         return Arrays.asList(anna, billy, steve, magda);
+    }
+    
+    public static void save(Employee employee){
+        try {
+            StringBuilder sb = new StringBuilder();
+            sb.append("### EMPLOYEE RECORD ####");
+            sb.append(System.lineSeparator());
+            sb.append("NAME: ");
+            sb.append(employee.getFullName());
+            sb.append(System.lineSeparator());
+            sb.append("POSITION: ");
+            sb.append(employee.getClass().getTypeName());
+            sb.append(System.lineSeparator());
+            sb.append("EMAIL: ");
+            sb.append(employee.getEmail());
+            sb.append(System.lineSeparator());
+            sb.append("MONTHLY WAGE: ");
+            sb.append(employee.getMonthlyIncome());
+            sb.append(System.lineSeparator());
+
+            Path path = Paths.get(employee.getFullName()
+                    .replace(" ","_") + ".rec");
+            Files.write(path, sb.toString().getBytes());
+
+            System.out.println("Saved employee " + employee.toString());
+        } catch (IOException e){
+            System.out.println("ERROR: Could not save employee. " + e);
+        }
     }
 }
